part 2

+++ src/main/java/net/id/paradiselost/blocks/ParadiseLostPortalBlock.java
@@ -25,23 +25,23 @@ public class ParadiseLostPortalBlock extends CustomPortalBlock {
     public void randomDisplayTick(BlockState state, World world, BlockPos pos, Random random) {
         if (random.nextInt(200) == 0) {
             world.getProfiler().push("portal");
-            world.playSound((double) pos.getX() + 0.5D, (double) pos.getY() + 0.5D, (double) pos.getZ() + 0.5D, ParadiseLostSoundEvents.BLOCK_PORTAL_AMBIENT, SoundCategory.BLOCKS, 0.5F, random.nextFloat() * 0.4F + 0.8F, false);
+            world.playSound(pos.getX() + 0.5D, pos.getY() + 0.5D, pos.getZ() + 0.5D, ParadiseLostSoundEvents.BLOCK_PORTAL_AMBIENT, SoundCategory.BLOCKS, 0.5F, random.nextFloat() * 0.4F + 0.8F, false);
             world.getProfiler().pop();
         }
 
-        double d = (double) pos.getX() + random.nextDouble();
-        double e = (double) pos.getY() + random.nextDouble();
-        double f = (double) pos.getZ() + random.nextDouble();
-        double g = ((double) random.nextFloat() - 0.5D) * 0.5D;
-        double h = ((double) random.nextFloat() - 0.5D) * 0.5D;
-        double j = ((double) random.nextFloat() - 0.5D) * 0.5D;
+        double d = pos.getX() + random.nextDouble();
+        double e = pos.getY() + random.nextDouble();
+        double f = pos.getZ() + random.nextDouble();
+        double g = (random.nextFloat() - 0.5D) * 0.5D;
+        double h = (random.nextFloat() - 0.5D) * 0.5D;
+        double j = (random.nextFloat() - 0.5D) * 0.5D;
         int k = random.nextInt(2) * 2 - 1;
         if (!world.getBlockState(pos.west()).isOf(this) && !world.getBlockState(pos.east()).isOf(this)) {
-            d = (double) pos.getX() + 0.5D + 0.25D * (double) k;
-            g = random.nextFloat() * 2.0F * (float) k;
+            d = pos.getX() + 0.5D + 0.25D * k;
+            g = random.nextFloat() * 2.0F * k;
         } else {
-            f = (double) pos.getZ() + 0.5D + 0.25D * (double) k;
-            j = random.nextFloat() * 2.0F * (float) k;
+            f = pos.getZ() + 0.5D + 0.25D * k;
+            j = random.nextFloat() * 2.0F * k;
         }
         if (world.getRandom().nextInt(6) != 0)
             world.addParticle(ParticleTypes.DRIPPING_WATER, d, e, f, g, h, j);
